requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.5.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-filter==1.1.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.4
packaging==23.2
passlib==1.7.4
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
